feat: add network interface capability
