+++ cmd/network/evpn/evpn-logical-brige.go
@@ -3,8 +3,8 @@
 // Copyright (c) 2022-2023 Dell Inc, or its subsidiaries.
 // Copyright (c) 2024 Ericsson AB.
 
-// Package network implements the network related CLI commands
-package network
+// Package evpn implements the evpn network related CLI commands
+package evpn
 
 import (
 	"context"
